chore: refactor project structure for consistency
- move scattered scripts to proper subdirectories (scripts/ts/, scripts/js/)
- fix relative import path in assign-permissions.ts after directory move
- add junction table comment header to prisma/models/relations.prisma
- move tencent-meeting API reference docs from src/ to docs/developer/
- rename src/task/dtos/ to src/task/dto/ to match project conventions
- update all 13 import paths in tasks.decorators.ts, tasks.service.ts, tasks.controller.ts
