+++ .gitignore
@@ -80,3 +80,4 @@ report.[0-9]*.[0-9]*.[0-9]*.[0-9]*.json
 .claude
 .kiro
 .trae
+# File Tree: nove_api.md
