+++ scripts/ts/test-oauth-setup.ts
