+++ src/task/tasks.controller.ts
@@ -23,10 +23,10 @@ import {
 import { ApiTags, ApiExtraModels } from '@nestjs/swagger';
 import { NoPermissionRequired } from '@/permission/decorators/permissions.decorator';
 import { TasksService } from './service/tasks.service';
-import { CreateOnceDto } from './dtos/create-once.dto';
-import { CreateCronDto } from './dtos/create-cron.dto';
-import { UpdateTaskDto } from './dtos/update-task.dto';
-import { QueryDto } from './dtos/query.dto';
+import { CreateOnceDto } from './dto/create-once.dto';
+import { CreateCronDto } from './dto/create-cron.dto';
+import { UpdateTaskDto } from './dto/update-task.dto';
+import { QueryDto } from './dto/query.dto';
 import {
   ApiHealthCheckDocs,
   ApiCreateOnceDocs,
@@ -46,7 +46,7 @@ import {
   RunNowResponse,
   TaskEntity,
   PaginatedTasksResponse,
-} from './dtos/responses.dto';
+} from './dto/responses.dto';
 
 @ApiTags('Tasks')
 @ApiExtraModels(TaskEntity, PaginatedTasksResponse, OkResponse, RunNowResponse)
